Splitting up test_fa_submission.py into separate files per class
